// src/types/replies.ts
export interface IReply {
    kind: Kind;
    data: IReplyData;
}

export interface IReplyData {
    children?:                       string;
    total_awards_received:           number;
    approved_at_utc:                 null;
    author_is_blocked:               boolean;
    comment_type:                    null;
    awarders:                        any[];
    mod_reason_by:                   null;
    banned_by:                       null;
    ups:                             number;
    removal_reason:                  null;
    link_id:                         LinkID;
    author_flair_template_id:        null;
    likes:                           null;
    replies:                         DataRepliesClass;
    user_reports:                    any[];
    saved:                           boolean;
    id:                              string;
    banned_at_utc:                   null;
    mod_reason_title:                null;
    gilded:                          number;
    archived:                        boolean;
    collapsed_reason_code:           string;
    no_follow:                       boolean;
    author:                          string;
    can_mod_post:                    boolean;
    send_replies:                    boolean;
    parent_id:                       string;
    score:                           number;
    approved_by:                     null;
    report_reasons:                  null;
    all_awardings:                   AllAwarding[];
    subreddit_id:                    SubredditID;
    body:                            string;
    edited:                          boolean;
    author_flair_css_class:          null;
    collapsed:                       boolean;
    downs:                           number;
    is_submitter:                    boolean;
    body_html:                       string;
    gildings:                        Gildings;
    collapsed_reason:                null;
    associated_award:                null;
    stickied:                        boolean;
    subreddit_type:                  SubredditType;
    can_gild:                        boolean;
    top_awarded_type:                null;
    unrepliable_reason:              null;
    author_flair_text_color:         string;
    score_hidden:                    boolean;
    permalink:                       string;
    num_reports:                     null;
    locked:                          boolean;
    name:                            Name;
    created:                         number;
    subreddit:                       Subreddit;
    author_flair_text:               null;
    treatment_tags:                  string[];
    created_utc:                     number;
    subreddit_name_prefixed:         SubredditNamePrefixed;
    controversiality:                number;
    depth:                           number;
    author_flair_background_color:   string;
    collapsed_because_crowd_control: null;
    mod_reports:                     any[];
    mod_note:                        null;
    distinguished:                   null;
}

export interface AllAwarding {
    giver_coin_reward:                    null;
    subreddit_id:                         null;
    is_new:                               boolean;
    days_of_drip_extension:               null;
    coin_price:                           number;
    id:                                   string;
    penny_donate:                         null;
    coin_reward:                          number;
    icon_url:                             string;
    days_of_premium:                      null;
    icon_height:                          number;
    tiers_by_required_awardings:          null;
    resized_icons:                        ResizedIcon[];
    icon_width:                           number;
    static_icon_width:                    number;
    start_date:                           null;
    is_enabled:                           boolean;
    awardings_required_to_grant_benefits: null;
    description:                          string;
    end_date:                             null;
    sticky_duration_seconds:              null;
    subreddit_coin_reward:                number;
    count:                                number;
    static_icon_height:                   number;
    name:                                 string;
    resized_static_icons:                 ResizedIcon[];
    icon_format:                          string;
    award_sub_type:                       string;
    penny_price:                          number;
    award_type:                           string;
    static_icon_url:                      string;
}

export interface ResizedIcon {
    url:    string;
    width:  number;
    height: number;
}

export interface Gildings {
}

export enum LinkID {
    T311Al04Z = "t3_11al04z",
}

export enum Name {
    T1J9Ssmjf = "t1_j9ssmjf",
}

export interface DataRepliesClass {
    kind: string;
    data: PurpleData;
}
// ...
export enum Subreddit {
    ThatsInsane = "ThatsInsane",
}

export enum SubredditID {
    T5Swxxz = "t5_swxxz",
}

export enum SubredditNamePrefixed {
    RThatsInsane = "r/ThatsInsane",
}

export enum SubredditType {
    Public = "public",
}

export enum Kind {
    T1 = "t1",
    more = 'more'
}